// src/app/models/users.ts
export class Users {
    id: string;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    location: string;
    otherDetails: string;
    roleName: string;
    displayPassword: string;
    password: string;
    email: string;
    isActive: boolean;
    filteredCount: number;
    termsAndConditonsAcceptedOn: Date
}